fixed avg and new value calculations

+++ include/ERG_Mode.h
@@ -105,7 +105,7 @@ class PowerTable {
   // Catalogs a new entry into the power table.
   void newEntry(PowerBuffer& powerBuffer);
 
-  void enterData(int i, int j, float pos); 
+  void enterData(int i, int j, int pos); 
 
   // returns incline for wattTarget. Null if not found.
   int32_t lookup(int watts, int cad);
